requirements.txt
gitpython
pooch
